Moves mocks to a specific mocks folder, and cleans typescript warnings.
